+++ controllers/getAllContacts.js
@@ -1,8 +1,8 @@
 const { Contact } = require("../models")
 
-const getAll = async (req, res) => {
+const getAllContacts = async (req, res) => {
     const contacts = await Contact.find();
     res.json(contacts)
 }
 
-module.exports = getAll;
+module.exports = getAllContacts;
