+++ models/index.js
@@ -1,5 +1,11 @@
-const { Contact } = require("./contact")
+const { Contact } = require("./contact");
+const { joiContactAddSchema } = require("./contact");
+const { joiContactUpdateSchema } = require("./contact");
+const { joiContactUpdateIsFavoriteSchema } = require("./contact");
 
 module.exports = {
-    Contact
+    Contact,
+    joiContactAddSchema,
+    joiContactUpdateSchema,
+    joiContactUpdateIsFavoriteSchema
 };
